yard object spawning (mostly) working

+++ src/world/Inventory/HousingItem.cpp
@@ -4,12 +4,12 @@ Sapphire::Inventory::HousingItem::HousingItem( uint64_t uId, uint32_t catalogId 
   Sapphire::Item( uId, catalogId, false )
 { }
 
-float Sapphire::Inventory::HousingItem::getRot() const
+uint16_t Sapphire::Inventory::HousingItem::getRot() const
 {
   return m_rotation;
 }
 
-void Sapphire::Inventory::HousingItem::setRot( float rot )
+void Sapphire::Inventory::HousingItem::setRot( uint16_t rot )
 {
   m_rotation = rot;
 }
